Supply an almost animated image to illustrate repeated addition
Instead of putting all the repeated addtions of a point in a single
image, use a series of images, where we show the next point, the
line through point and basepoint, the third intersection with the
curve, and an arrow to mirror the point at the x-axis.

Including these images consecutively looks like an animation with
the additional advantage that the presenter can control the speed.